+++ services/advisor/mgmt/2020-01-01/advisor/advisorapi/interfaces.go
@@ -15,7 +15,7 @@ import (
 
 // RecommendationMetadataClientAPI contains the set of methods on the RecommendationMetadataClient type.
 type RecommendationMetadataClientAPI interface {
-	Get(ctx context.Context, name string) (result advisor.SetObject, err error)
+	Get(ctx context.Context, name string) (result advisor.MetadataEntity, err error)
 	List(ctx context.Context) (result advisor.MetadataEntityListResultPage, err error)
 	ListComplete(ctx context.Context) (result advisor.MetadataEntityListResultIterator, err error)
 }
@@ -56,7 +56,7 @@ var _ OperationsClientAPI = (*advisor.OperationsClient)(nil)
 type SuppressionsClientAPI interface {
 	Create(ctx context.Context, resourceURI string, recommendationID string, name string, suppressionContract advisor.SuppressionContract) (result advisor.SuppressionContract, err error)
 	Delete(ctx context.Context, resourceURI string, recommendationID string, name string) (result autorest.Response, err error)
-	Get(ctx context.Context, resourceURI string, recommendationID string, name string) (result advisor.SetObject, err error)
+	Get(ctx context.Context, resourceURI string, recommendationID string, name string) (result advisor.SuppressionContract, err error)
 	List(ctx context.Context, top *int32, skipToken string) (result advisor.SuppressionContractListResultPage, err error)
 	ListComplete(ctx context.Context, top *int32, skipToken string) (result advisor.SuppressionContractListResultIterator, err error)
 }