+++ services/advisor/mgmt/2020-01-01/advisor/CHANGELOG.md
@@ -1,7 +1,52 @@
-# Change History
+# Unreleased
+
+## Breaking Changes
+
+### Struct Changes
+
+#### Removed Structs
+
+1. SetObject
+
+### Signature Changes
+
+#### Funcs
+
+1. RecommendationMetadataClient.Get
+	- Returns
+		- From: SetObject, error
+		- To: MetadataEntity, error
+1. RecommendationMetadataClient.GetResponder
+	- Returns
+		- From: SetObject, error
+		- To: MetadataEntity, error
+1. SuppressionsClient.Get
+	- Returns
+		- From: SetObject, error
+		- To: SuppressionContract, error
+1. SuppressionsClient.GetResponder
+	- Returns
+		- From: SetObject, error
+		- To: SuppressionContract, error
 
 ## Additive Changes
 
 ### New Funcs
 
-1. Resource.MarshalJSON() ([]byte, error)
+1. ResourceMetadata.MarshalJSON() ([]byte, error)
+
+### Struct Changes
+
+#### New Struct Fields
+
+1. MetadataEntity.autorest.Response
+1. RecommendationProperties.Actions
+1. RecommendationProperties.Description
+1. RecommendationProperties.ExposedMetadataProperties
+1. RecommendationProperties.Label
+1. RecommendationProperties.LearnMoreLink
+1. RecommendationProperties.PotentialBenefits
+1. RecommendationProperties.Remediation
+1. ResourceMetadata.Action
+1. ResourceMetadata.Plural
+1. ResourceMetadata.Singular